Remove XACML related dependencies and tests
