+++ modules/api-resources/api-resources-full/pom.xml
@@ -442,10 +442,6 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.api.server.entitlement</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.framework</groupId>
             <artifactId>org.wso2.carbon.identity.api.server.configuration.mgt</artifactId>
         </dependency>
         <dependency>
