+++ modules/integration/tests-common/pom.xml
@@ -32,7 +32,6 @@
         <module>admin-clients</module>
         <module>ui-pages</module>
         <module>integration-test-utils</module>
-        <module>extensions</module>
         <module>jacoco-report-generator</module>
     </modules>
